# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "個人用家計簿アプリのバックエンドAPI（銀行CSVインポート・カテゴリ分類）",
        "title": "MoneyRabbit API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/import/confirm": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "CSVインポート確定（分類＋保存）",
                "parameters": [
                    {
                        "description": "インポートデータ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.confirmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.importResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/import/preview": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "CSVプレビュー（DB書き込みなし）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "フォーマットID（例: smbc_bank）",
                        "name": "import_format_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "CSVファイル",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.previewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "取引一覧",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "年（例: 2025）",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "月（1-12）",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "カテゴリID（UUID）",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ページ番号（0始まり）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ページサイズ（デフォルト: 50）",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.transactionListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{id}/category": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "取引のカテゴリを変更",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category ID（nullでクリア）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateTransactionCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Category": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CategoryRule"
                    }
                },
                "sort_order": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/entity.CategoryType"
                }
            }
        },
        "entity.CategoryRule": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "keyword": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
        "entity.CategoryType": {
            "type": "string",
            "enum": [
                "income",
                "expense",
                "both"
            ],
            "x-enum-varnames": [
                "CategoryTypeIncome",
                "CategoryTypeExpense",
                "CategoryTypeBoth"
            ]
        },
        "entity.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "category": {
                    "$ref": "#/definitions/entity.Category"
                },
                "category_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "import_format_id": {
                    "type": "string"
                },
                "imported_at": {
                    "type": "string"
                }
            }
        },
        "handler.confirmRequest": {
            "type": "object",
            "required": [
                "import_format_id",
                "transactions"
            ],
            "properties": {
                "import_format_id": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.txRowInput"
                    }
                }
            }
        },
...
        "handler.importResultResponse": {
            "type": "object",
            "properties": {
                "Errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "Imported": {
                    "type": "integer"
                },
                "Skipped": {
                    "type": "integer"
                }
            }
        },
        "handler.previewResponse": {
            "type": "object",
            "properties": {
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.rowResponse"
                    }
                },
                "skipped_rows": {
                    "type": "integer"
                }
            }
        },
        "handler.rowResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "handler.transactionListResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Transaction"
                    }
                }
            }
        },
        "handler.txRowInput": {
            "type": "object",
            "required": [
                "date"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
        "handler.updateTransactionCategoryRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                }
            }
        }
    }
}